Create initial algorithm
Water flows from the top of the terrain to the bottom.
At each step, the water will choose the lowest square beside the water to flood.
The highest point is purple and the lowest is orange.
The terrain is shown by higher places being lighter.
Definitely didn't steal Minecraft colors.(I didn't)