I added more terrain smoothing.
The terrain is smoothed by taking the average of each cell and its neighbors and setting that as the value of the cell.
Now I added a for loop to do this multiple times so the terrain is smoother. Adjust smoothing_passes to change smoothness. With high smoothness ( > 3 ) the water tends to form large lakes.